# public/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This API manages routes for public transport agencies. Saving your data in a relational database, converts the information to the GTFS specification, which can be used to search for information from traffic agencies and to monitor in real time any specific itinerary. \n\nImplemented for a graduate work in the Computer Science course. \n\n It is necessary to have configured the SMTP server credentials to notify the user via e-mail with the generated password."
  version: "1.0.0"
  title: "GTFS Manager"
  contact:
    email: "matheusinhesta@live.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api"
...
paths:
...
  /forget-password:
    post:
      tags:
        - "Auth"
      summary: "Send a new password to the user's e-mail"
      description: "It is necessary to have configured the SMTP server credentials to notify the user via e-mail with the generated a new password"
      operationId: "forgetPassword"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "email"
          in: "formData"
          description: "The user e-mail address for login"
          required: true
          type: "string"

      responses:
        "200":
          description: "Success! An e-mail was sent to the registered address with the user generated a new password."
        "400":
          description: "Invalid input."
        "404":
          description: "User not found."
        "500":
          description: "The token could not be created"
  /change-password:
    post:
      tags:
        - "Auth"
      summary: "Change the password of the authenticated user"
      description: ""
      operationId: "changePassword"
      consumes:
        - "application/x-www-form-urlencoded"
      produces:
        - "application/json"
      parameters:
        - name: "Authorization"
          in: "header"
          description: "Access token"
          required: true
          type: "string"
          format: "Bearer JWT token"

        - name: "password"
          in: "formData"
          description: "The user password"
          required: true
          type: "string"

        - name: "new_password"
          in: "formData"
          description: "The new password"
          required: true
          type: "string"

        - name: "new_password_confirmation"
          in: "formData"
          description: "The new password confirmation"
          required: true
          type: "string"

      responses:
        "200":
          description: "Success! User password has been changed."
        "400":
          description: "Invalid input."
        "500":
          description: "The new password could not be updated."
  /refresh:
    get:
      tags:
        - "Auth"
      summary: "Refresh access token"
      description: ""
      operationId: "refresh"
      produces:
        - "application/json"
      parameters:
        - name: "Authorization"
          in: "header"
          description: "Access token"
          required: true
          type: "string"
          format: "Bearer JWT token"
      responses:
        "200":
          description: "Successful operation"
          headers:
            access_token:
              type: "string"
              description: "JWT token used for any request that needs to be authenticated"
            token_type:
              type: "string"
              description: "Tokens type"
            expires_in:
              type: "string"
              format: "date-time"
              description: "Time in UTC when token expires"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
...